:rotating_light: lint fix

+++ contracts/interfaces/IUniswapV2Pair.sol
@@ -1,12 +1,8 @@
 pragma solidity =0.8.10;
 
 interface IUniswapV2Pair {
-    event Approval(
-        address indexed owner,
-        address indexed spender,
-        uint256 value
-    );
-    event Transfer(address indexed from, address indexed to, uint256 value);
+    event Approval(address indexed owner, address indexed spender, uint value);
+    event Transfer(address indexed from, address indexed to, uint value);
 
     function name() external pure returns (string memory);
 
@@ -14,59 +10,59 @@ interface IUniswapV2Pair {
 
     function decimals() external pure returns (uint8);
 
-    function totalSupply() external view returns (uint256);
+    function totalSupply() external view returns (uint);
 
-    function balanceOf(address owner) external view returns (uint256);
+    function balanceOf(address owner) external view returns (uint);
 
     function allowance(address owner, address spender)
         external
         view
-        returns (uint256);
+        returns (uint);
 
-    function approve(address spender, uint256 value) external returns (bool);
+    function approve(address spender, uint value) external returns (bool);
 
-    function transfer(address to, uint256 value) external returns (bool);
+    function transfer(address to, uint value) external returns (bool);
 
     function transferFrom(
         address from,
         address to,
-        uint256 value
+        uint value
     ) external returns (bool);
 
     function DOMAIN_SEPARATOR() external view returns (bytes32);
 
     function PERMIT_TYPEHASH() external pure returns (bytes32);
 
-    function nonces(address owner) external view returns (uint256);
+    function nonces(address owner) external view returns (uint);
 
     function permit(
         address owner,
         address spender,
-        uint256 value,
-        uint256 deadline,
+        uint value,
+        uint deadline,
         uint8 v,
         bytes32 r,
         bytes32 s
     ) external;
 
-    event Mint(address indexed sender, uint256 amount0, uint256 amount1);
+    event Mint(address indexed sender, uint amount0, uint amount1);
     event Burn(
         address indexed sender,
-        uint256 amount0,
-        uint256 amount1,
+        uint amount0,
+        uint amount1,
         address indexed to
     );
     event Swap(
         address indexed sender,
-        uint256 amount0In,
-        uint256 amount1In,
-        uint256 amount0Out,
-        uint256 amount1Out,
+        uint amount0In,
+        uint amount1In,
+        uint amount0Out,
+        uint amount1Out,
         address indexed to
     );
     event Sync(uint112 reserve0, uint112 reserve1);
 
-    function MINIMUM_LIQUIDITY() external pure returns (uint256);
+    function MINIMUM_LIQUIDITY() external pure returns (uint);
 
     function factory() external view returns (address);
 
@@ -83,21 +79,19 @@ interface IUniswapV2Pair {
             uint32 blockTimestampLast
         );
 
-    function price0CumulativeLast() external view returns (uint256);
+    function price0CumulativeLast() external view returns (uint);
 
-    function price1CumulativeLast() external view returns (uint256);
+    function price1CumulativeLast() external view returns (uint);
 
-    function kLast() external view returns (uint256);
+    function kLast() external view returns (uint);
 
-    function mint(address to) external returns (uint256 liquidity);
+    function mint(address to) external returns (uint liquidity);
 
-    function burn(address to)
-        external
-        returns (uint256 amount0, uint256 amount1);
+    function burn(address to) external returns (uint amount0, uint amount1);
 
     function swap(
-        uint256 amount0Out,
-        uint256 amount1Out,
+        uint amount0Out,
+        uint amount1Out,
         address to,
         bytes calldata data
     ) external;
@@ -110,20 +104,20 @@ interface IUniswapV2Pair {
 }
 
 interface IUniswapV2Router {
-    function getAmountsOut(uint256 amountIn, address[] memory path)
+    function getAmountsOut(uint amountIn, address[] memory path)
         external
         view
-        returns (uint256[] memory amounts);
+        returns (uint[] memory amounts);
 
     //    function getAmountOut(uint amountIn, uint reserveIn, uint reserveOut) external pure returns (uint amountOut);
-    function getAmountsIn(uint256 amountOut, address[] memory path)
+    function getAmountsIn(uint amountOut, address[] memory path)
         external
         view
-        returns (uint256[] memory amounts);
+        returns (uint[] memory amounts);
 
     function quote(
-        uint256 amountA,
-        uint256 reserveA,
-        uint256 reserveB
-    ) external pure returns (uint256 amountB);
+        uint amountA,
+        uint reserveA,
+        uint reserveB
+    ) external pure returns (uint amountB);
 }
